Methods

Methods

This section describes systems methods implemented in MathematicalSets.jl.

Methods

Metric properties

Topological properties

Metric properties

MathematicalSets.surface_area — Function.

surface_area(S::AbstractSet)

Return the surface area of the given set.

Input

S – set

Output

A number.

Notes

The surface_area refers to the surface area of the set with respect to some measure. The actual notion of surface adopted is left to the package implementing this function.

source

MathematicalSets.volume — Function.

volume(S::AbstractSet)

Return the volume of the given set.

Input

S – set

Output

A number.

Notes

The volume refers to the volume of the set with respect to some measure. The actual notion of volume adopted is left to the package implementing this function.

source

Topological properties

MathematicalSets.dimension — Function.

dimension(S::AbstractSet)

Return the dimension of the given set.

Input

S – set

Output

An integer.

Notes

The actual notion of dimension adopted is left to the package implementing this function.

For instance, in many applications dimension refers to the dimension of the affine hull of the set. For example, a line in an $n$-dimensional space has dimension equal to $1$, even if the dimension of the space is $n$. However, in other domains, e.g. semialgebraic sets, dimension is not the dimension of its affine hull.

See also space_dimension(::AbstractSet) for the ambient dimension of the set.

source

MathematicalSets.space_dimension — Function.

space_dimension(S::AbstractSet)

Return the space dimension of the given set.

Input

S – set

Output

An integer.

Notes

The space_dimension refers to the ambient dimension of the set. For example, a line in an $n$-dimensional space has space_dimension equal to $n$, even if the dimension of its affine hull is $1$.

See also dimension(::AbstractSet) for the dimension of the set.

source